updates on language menu
