+++ src/App.js
@@ -8,6 +8,7 @@ import WhatWeDo from "./pages/WhatWeDo";
 import Languages from "./pages/Languages";
 import Backdrop from "./components/Backdrop";
 import { StateContext } from "./store/StateContext";
+import LanguageMenu from "./components/LanguageMenu";
 
 function App() {
   const [showArrow, setShowArrow] = useState(false);
@@ -102,6 +103,7 @@ function App() {
   return (
     <StateContext>
       <div className="App">
+        <LanguageMenu />
         {showMenu ? <Backdrop onClick={backDropClickHandler} /> : ""}
         {showArrow && upIcon}
         <Home showMenu={showMenu} setShowMenu={setShowMenu} />
