+++ src/components/Nav.js
@@ -1,15 +1,10 @@
 import React from "react";
-import bgFlagEmoji from "../assets/bg-flag-emoji.png";
-import enFlagEmoji from "../assets/en-flag-emoji.png";
+import { languages } from "../languages/languages";
 import styles from "./Nav.module.css";
 import { useState, useEffect, useRef } from "react";
 import logo from "../assets/logo151.png";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faNavicon,
-  faClose,
-  faCaretDown,
-} from "@fortawesome/free-solid-svg-icons";
+import { faNavicon, faClose } from "@fortawesome/free-solid-svg-icons";
 import { withNamespaces } from "react-i18next";
 import i18n from "../i18n";
 import {
@@ -22,18 +17,6 @@ const Nav = (props) => {
   const { t } = props;
   const navRef = useRef();
   const [isDesktop, setDesktop] = useState(window.innerWidth > 650);
-  const [selectedLanguage, setSelectedLanguage] = useState(bgFlagEmoji);
-  const [dropDown, setDropDown] = useState(false);
-
-  const languageChangeDropHandler = () => {
-    setDropDown((prevState) => !prevState);
-  };
-
-  const languageChangeHandler = (e) => {
-    i18n.changeLanguage(e.target.id);
-    let language = document.getElementById(e.target.id);
-    setSelectedLanguage(language.firstChild.src);
-  };
 
   const closeIcon = (
     <FontAwesomeIcon
@@ -52,7 +35,6 @@ const Nav = (props) => {
       />
     </div>
   );
-  const downIcon = <FontAwesomeIcon icon={faCaretDown}></FontAwesomeIcon>;
   const updateMedia = () => {
     setDesktop(window.innerWidth > 576);
   };
@@ -108,45 +90,6 @@ const Nav = (props) => {
             <a onClick={iconClickHandler} href="#GetInTouch">
               {t("NavigationBarContactUs")}
             </a>
-          </div>
-          <div
-            onClick={languageChangeDropHandler}
-            className={styles.customSelect}
-          >
-            <div className={styles.currentlySelectedContainer}>
-              <img
-                className={styles.currentlySelected}
-                src={selectedLanguage}
-                alt=""
-              />
-              {downIcon}
-            </div>
-            <div
-              style={dropDown ? { display: "block" } : {}}
-              className={styles.dropDown}
-            >
-              <div
-                onClick={languageChangeHandler}
-                id="bg"
-                className={styles.dropDownItem}
-              >
-                <img
-                  className={styles.flagEmoji}
-                  id="bg"
-                  src={bgFlagEmoji}
-                  alt=""
-                />
-                <p id="bg">Bulgarian</p>
-              </div>
-              <div
-                onClick={languageChangeHandler}
-                id="en"
-                className={styles.dropDownItem}
-              >
-                <img className={styles.flagEmoji} src={enFlagEmoji} alt="" />
-                <p id="en">English</p>
-              </div>
-            </div>
           </div>
         </nav>
         {!isDesktop && navIcon}
