+++ src/components/Nav.module.css
@@ -15,6 +15,7 @@
 }
 
 nav {
+  margin-right: 6rem;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -67,65 +68,7 @@ nav a {
   margin-inline: auto;
   background-color: #303030;
 }
-.customSelect {
-  cursor: pointer;
-  position: relative;
-}
-.currentlySelectedContainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-right: 2rem;
-}
 
-.currentlySelected {
-  height: 30px;
-  margin-right: 20px;
-  cursor: pointer;
-}
-.dropDown {
-  padding: 0.5rem;
-  background-color: white;
-  display: none;
-  position: absolute;
-  top: 50px;
-  right: 13px;
-  border-radius: 5px;
-}
-.dropDown::after {
-  background-color: white;
-  position: absolute;
-  content: "";
-  height: 10px;
-  width: 10px;
-  top: -10px;
-  right: 15px;
-  width: 0;
-  height: 0;
-  border-style: solid;
-  border-width: 0 10px 10px 10px;
-  border-color: #bad9ed #bad9ed white #bad9ed;
-}
-
-.dropDownItem {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: fit-content;
-  height: fit-content;
-  cursor: pointer;
-}
-.dropDownItem span {
-  display: inline;
-  height: fit-content;
-}
-.dropDown p {
-  margin: 0px;
-}
-
-.flagEmoji {
-  height: 30px;
-}
 /*
 /* Smartphone devices (576px and down) */
 @media (max-width: 576px) {
@@ -184,6 +127,15 @@ nav a {
     width: var(--logo-size-xs);
     height: var(--logo-size-xs);
   }
+  .customSelectMobile {
+    position: fixed;
+    top: 30px;
+    right: 10px;
+  }
+
+  .dropDown {
+    right: 0.7rem;
+  }
 }
 
 /* Small devices (tablets 576px-768px) */
@@ -197,6 +149,9 @@ nav a {
   }
   nav a {
     font-size: var(--font-size-xs);
+  }
+  .dropDown {
+    right: 0.7rem;
   }
 }
