+++ src/languages/languages.js
@@ -1,0 +1,7 @@
+import bgFlagEmoji from "../assets/bg-flag-emoji.png";
+import enFlagEmoji from "../assets/en-flag-emoji.png";
+
+export const languages = [
+  { language: "Български", languageKey: "bg", image: bgFlagEmoji },
+  { language: "English", languageKey: "en", image: enFlagEmoji },
+];
